requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.6.1,<3.0.0
python-dotenv==1.0.1
pytest==8.0.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.12.0
langchain-community==0.0.19
langchain-openai==0.0.5
python-dateutil==2.8.2
sqlalchemy==2.0.27 